backend/requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
openai
aiohttp
sounddevice
scipy
librosa
numpy
pydantic
python-dotenv
